requirements.txt
openai>=1.3.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
tiktoken>=0.4.0
tenacity>=8.0.0
pyyaml>=6.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
vaderSentiment>=3.3.0
textblob>=0.17.0